Major clean-up: - Goodbye Main and XmlvmBuilder. - Removing dependency on czUtil by using the new UniversalFile API.
git-svn-id: http://xmlvm.svn.sourceforge.net/svnroot/xmlvm/trunk/xmlvm@948 0874c2aa-48f6-4a0f-8a6b-e32f4465a257

+++ src/xmlvm/org/xmlvm/util/universalfile/UniversalFile.java
@@ -129,6 +129,11 @@ public abstract class UniversalFile {
      */
     public boolean saveFileAs(String path) {
         try {
+            // Make sure the destination directory exists.
+            File parent = (new File(path)).getParentFile();
+            if (!parent.exists() && !parent.mkdirs()) {
+                return false;
+            }
             FileOutputStream outputStream = new FileOutputStream(path);
             outputStream.write(getFileAsBytes());
             outputStream.close();
